Run migration in test mode too

+++ test/test_helper.rb
@@ -60,8 +60,8 @@ class ActiveSupport::TestCase
 
   def setup
     CheckNotifications::Slack::Request.any_instance.stubs(:request).returns(nil)
-    [Annotation, Team, TeamUser].each{ |klass| klass.delete_all }
-    [ProjectMedia, Media, Account, Source, User, Annotation, DynamicAnnotation::AnnotationType, DynamicAnnotation::FieldType, DynamicAnnotation::FieldInstance].each{ |m| m.destroy_all }
+    [Annotation, Team, TeamUser, DynamicAnnotation::AnnotationType, DynamicAnnotation::FieldType, DynamicAnnotation::FieldInstance].each{ |klass| klass.delete_all }
+    [ProjectMedia, Media, Account, Source, User, Annotation].each{ |m| m.destroy_all }
     # create index
     MediaSearch.delete_index
     MediaSearch.create_index
